+++ BlackJack/model/rules/AmericanNewGameStrategy.cs
@@ -1,31 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
+﻿
 namespace BlackJack.model.rules
 {
     class AmericanNewGameStrategy : INewGameStrategy
     {
         public bool NewGame(Deck a_deck, Dealer a_dealer, Player a_player)
         {
-            Card c;
-
-            c = a_deck.GetCard();
-            c.Show(true);
-            a_player.DealCard(c);
-
-            c = a_deck.GetCard();
-            c.Show(true);
-            a_dealer.DealCard(c);
-
-            c = a_deck.GetCard();
-            c.Show(true);
-            a_player.DealCard(c);
-
-            c = a_deck.GetCard();
-            c.Show(false);
-            a_dealer.DealCard(c);
+            a_dealer.DealCard(a_player, true);
+            a_dealer.DealCard(a_dealer, true);
+            a_dealer.DealCard(a_player, true);
+            a_dealer.DealCard(a_dealer, false);
 
             return true;
         }