+++ BlackJack/model/rules/InternationalNewGameStrategy.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
+﻿
 namespace BlackJack.model.rules
 {
     class InternationalNewGameStrategy : INewGameStrategy
@@ -10,19 +6,9 @@ namespace BlackJack.model.rules
 
         public bool NewGame(Deck a_deck, Dealer a_dealer, Player a_player)
         {
-            Card c;
-
-            c = a_deck.GetCard();
-            c.Show(true);
-            a_player.DealCard(c);
-
-            c = a_deck.GetCard();
-            c.Show(true);
-            a_dealer.DealCard(c);
-
-            c = a_deck.GetCard();
-            c.Show(true);
-            a_player.DealCard(c);
+            a_dealer.DealCard(a_player, true);
+            a_dealer.DealCard(a_dealer, true);
+            a_dealer.DealCard(a_player, true);
 
             return true;
         }